API, Frontend - Add server-side error handling
This change adds server-side checks against supplied values for Name, Age, and Title. If any check fails, the client is redirected to the new error.html page.

It is worth noting that in testing, the only way to reach this page is if the user modifies the webpage content in the browser to remove the client-side form validation